added book rating

+++ FoxBook-app/app/src/main/java/com/example/foxbook/api/UserRating.kt
@@ -1,3 +1,3 @@
 package com.example.foxbook.api
 
-class UserRating(val user_rating: Int)
+class UserRating(val user_rating: Float)
